requirements.txt
asgiref==3.8.1
Django==4.2.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==33.3.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
